Move StorageActionsTask to Storage RP
