+++ internal/clients/client.go
@@ -138,7 +138,6 @@ import (
 	signalr "github.com/hashicorp/terraform-provider-azurerm/internal/services/signalr/client"
 	appPlatform "github.com/hashicorp/terraform-provider-azurerm/internal/services/springcloud/client"
 	storage "github.com/hashicorp/terraform-provider-azurerm/internal/services/storage/client"
-	storageActions "github.com/hashicorp/terraform-provider-azurerm/internal/services/storageactions/client"
 	storageCache "github.com/hashicorp/terraform-provider-azurerm/internal/services/storagecache/client"
 	storageMover "github.com/hashicorp/terraform-provider-azurerm/internal/services/storagemover/client"
 	streamAnalytics "github.com/hashicorp/terraform-provider-azurerm/internal/services/streamanalytics/client"
@@ -276,7 +275,6 @@ type Client struct {
 	ServiceNetworking                 *servicenetworking_2025_01_01.Client
 	SignalR                           *signalr.Client
 	Storage                           *storage.Client
-	StorageActions                    *storageActions.Client
 	StorageCache                      *storagecache_2024_07_01.Client
 	StorageCache_2023_05_01           *storagecache_2023_05_01.Client
 	StorageMover                      *storageMover.Client
@@ -647,9 +645,6 @@ func (client *Client) Build(ctx context.Context, o *common.ClientOptions) error 
 	}
 	if client.Storage, err = storage.NewClient(o); err != nil {
 		return fmt.Errorf("building clients for Storage: %+v", err)
-	}
-	if client.StorageActions, err = storageActions.NewClient(o); err != nil {
-		return fmt.Errorf("building clients for StorageActions: %+v", err)
 	}
 	if client.StorageCache, err = storageCache.NewClient(o); err != nil {
 		return fmt.Errorf("building clients for Storage Cache: %+v", err)
