+++ internal/provider/services.go
@@ -124,7 +124,6 @@ import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/signalr"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/springcloud"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/storage"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/services/storageactions"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/storagecache"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/storagemover"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/services/streamanalytics"
@@ -229,7 +228,6 @@ func SupportedTypedServices() []sdk.TypedServiceRegistration {
 		signalr.Registration{},
 		springcloud.Registration{},
 		storage.Registration{},
-		storageactions.Registration{},
 		storagecache.Registration{},
 		storagemover.Registration{},
 		streamanalytics.Registration{},
@@ -468,7 +466,6 @@ func SupportedFrameworkServices() []sdk.FrameworkServiceRegistration {
 		signalr.Registration{},
 		springcloud.Registration{},
 		storage.Registration{},
-		storageactions.Registration{},
 		storagecache.Registration{},
 		storagemover.Registration{},
 		streamanalytics.Registration{},
