+++ internal/services/storage/client/client.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	storage "github.com/hashicorp/go-azure-sdk/resource-manager/storage/2025-08-01"
+	"github.com/hashicorp/go-azure-sdk/resource-manager/storageactions/2023-01-01/storagetasks"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/storagesync/2020-03-01/cloudendpointresource"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/storagesync/2020-03-01/registeredserverresource"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/storagesync/2020-03-01/serverendpointresource"
@@ -30,6 +31,8 @@ type Client struct {
 	SyncRegisteredServerClient *registeredserverresource.RegisteredServerResourceClient
 	SyncServerEndpointsClient  *serverendpointresource.ServerEndpointResourceClient
 	SyncServiceClient          *storagesyncservicesresource.StorageSyncServicesResourceClient
+
+	StorageTasksClient *storagetasks.StorageTasksClient
 
 	authConfigForAzureAD *auth.Credentials
 }
@@ -80,6 +83,12 @@ func NewClient(o *common.ClientOptions) (*Client, error) {
 	}
 	o.Configure(syncGroupsClient.Client, o.Authorizers.ResourceManager)
 
+	storageTasksClient, err := storagetasks.NewStorageTasksClientWithBaseURI(o.Environment.ResourceManager)
+	if err != nil {
+		return nil, fmt.Errorf("building StorageTasks client: %+v", err)
+	}
+	o.Configure(storageTasksClient.Client, o.Authorizers.ResourceManager)
+
 	// TODO: switch Storage Containers to using the storage.BlobContainersClient
 	// (which should fix #2977) when the storage clients have been moved in here
 	client := Client{
@@ -89,6 +98,8 @@ func NewClient(o *common.ClientOptions) (*Client, error) {
 		SyncServerEndpointsClient:  syncServerEndpointClient,
 		SyncServiceClient:          syncServiceClient,
 		SyncGroupsClient:           syncGroupsClient,
+
+		StorageTasksClient: storageTasksClient,
 
 		StorageDomainSuffix: *storageSuffix,
 	}
