--- internal/services/storageactions/registration.go
@@ -1,63 +1,0 @@
-// Copyright IBM Corp. 2014, 2025
-// SPDX-License-Identifier: MPL-2.0
-
-package storageactions
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/action"
-	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
-	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
-)
-
-type Registration struct{}
-
-var (
-	_ sdk.FrameworkServiceRegistration = Registration{}
-	_ sdk.TypedServiceRegistration     = Registration{}
-)
-
-func (r Registration) AssociatedGitHubLabel() string {
-	return "service/storageactions"
-}
-
-// Name is the name of this Service
-func (r Registration) Name() string {
-	return "Storage Actions"
-}
-
-// WebsiteCategories returns a list of categories which can be used for the sidebar
-func (r Registration) WebsiteCategories() []string {
-	return []string{
-		"Storage Actions",
-	}
-}
-
-func (r Registration) DataSources() []sdk.DataSource {
-	return []sdk.DataSource{}
-}
-
-func (r Registration) Resources() []sdk.Resource {
-	return []sdk.Resource{
-		StorageActionsTaskResource{},
-	}
-}
-
-func (r Registration) Actions() []func() action.Action {
-	return []func() action.Action{}
-}
-
-func (r Registration) FrameworkResources() []sdk.FrameworkWrappedResource {
-	return []sdk.FrameworkWrappedResource{}
-}
-
-func (r Registration) FrameworkDataSources() []sdk.FrameworkWrappedDataSource {
-	return []sdk.FrameworkWrappedDataSource{}
-}
-
-func (r Registration) EphemeralResources() []func() ephemeral.EphemeralResource {
-	return []func() ephemeral.EphemeralResource{}
-}
-
-func (r Registration) ListResources() []sdk.FrameworkListWrappedResource {
-	return []sdk.FrameworkListWrappedResource{}
-}
